circleci: Use old build folders if they exist
In install scripts, if the folder already exists, reuse it - don't
download it again.

+++ .misc/install.python-dbus.sh
@@ -1,13 +1,17 @@
 set -x
 set -e
 
-echo Downloading python-dbus...
-wget http://dbus.freedesktop.org/releases/dbus-python/dbus-python-1.2.0.tar.gz -O dbus-python.tar.gz -q
-echo Unpacking python-dbus...
-tar -zxvf dbus-python.tar.gz > /dev/null
-rm dbus-python.tar.gz
-
-cd dbus-python-1.2.0
+if [ ! -d "dbus-python-1.2.0" ] ; then
+  echo Downloading python-dbus...
+  wget http://dbus.freedesktop.org/releases/dbus-python/dbus-python-1.2.0.tar.gz -O dbus-python.tar.gz -q
+  echo Unpacking python-dbus...
+  tar -zxvf dbus-python.tar.gz > /dev/null
+  rm dbus-python.tar.gz
+  cd dbus-python-1.2.0
+else
+  cd dbus-python-1.2.0
+  make clean >/dev/null || make clean
+fi
 
 ./configure --prefix=$python_virtualenv >/dev/null || ./configure --prefix=$python_virtualenv
 make >/dev/null || make
